+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Explosive.java
@@ -1,11 +1,14 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.damage.BonusDamageEnchantment;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.event.entity.living.LivingDeathEvent;
 
 public class Explosive extends BonusDamageEnchantment {
     public Explosive() {
@@ -23,15 +26,16 @@ public class Explosive extends BonusDamageEnchantment {
     }
 
     public static float getExplosionPower(int lvl, ItemStack stack) {
-        return Math.max(1, (1 + (lvl - 1) * 0.5f) * getDamageBonusRatio(stack));
+        return 1 + ((lvl - 1) * 0.5f);
     }
 
-    @Override
-    public void doPostAttack(LivingEntity attacker, Entity entity, int lvl) {
-        if (!(entity instanceof LivingEntity livingEntity))
+    public static void onKill(LivingDeathEvent event) {
+        if (!(event.getSource().getEntity() instanceof LivingEntity killer))
+            return;
+        int lvl = EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.EXPLOSIVE.get(), killer);
+        if (lvl <= 0)
             return;
 
-        if (lvl > 0)
-            livingEntity.level().explode(null, livingEntity.getX(), livingEntity.getY() + livingEntity.getBbHeight() / 2, livingEntity.getZ(), getExplosionPower(lvl, attacker.getMainHandItem()), Level.ExplosionInteraction.TNT);
+        event.getEntity().level().explode(event.getEntity(), event.getEntity().getX(), event.getEntity().getY(), event.getEntity().getZ(), getExplosionPower(lvl, killer.getMainHandItem()), Level.ExplosionInteraction.BLOCK);
     }
 }